+++ src/bin/rhai-run.rs
@@ -85,7 +85,7 @@ fn main() {
             .compile(contents)
             .map_err(|err| Box::new(err.into()) as Box<EvalAltResult>)
             .and_then(|mut ast| {
-                ast.set_source(filename.to_string_lossy());
+                ast.set_source(filename.to_string_lossy().to_string());
                 engine.consume_ast(&ast)
             })
         {